+++ reporting.py
@@ -53,10 +53,32 @@ class report:
 
         for file_name in file_list:
             self.console_out(f"Processing file: {file_name}")
-            self.write_entropy_graph(path.join(in_path, file_name), path.join(in_path, file_name) + ".csv", step)
+            self.write_file_graph(path.join(in_path, file_name), path.join(in_path, file_name) + ".csv", step)
 
+
+    def write_file_report(self, in_file_name, out_file_name):
+        """
+        Creates an entropy report for a single file and saves it in a CSV spreadsheet.
+        """
+        byte_array = []
+        with open(in_file_name, "rb") as f:
+            byte_array = f.read()
+
+        file_size = len(byte_array)
+        entropy = compute_entropy(byte_array)
+        new_pi = compute_monte_carlo_pi(byte_array)
+        pi_deviation = get_pi_deviation(new_pi)
+        chi_squared = compute_chi_squared(byte_array)
+
+        if out_file_name != "":
+            with open(out_file_name, "w") as file_out:
+                file_out.write("File Name;File Size (B);Entropy;Pi;Pi Deviation (%);Chi-Squared\r")
+                file_out.write("%s;%d;%f;%f;%f;%f\r" % (in_file_name, file_size, entropy, new_pi, pi_deviation, chi_squared))
+            
+        self.console_out("%30s %15s %15s %15s %15s %15s" % ("File Name", "File Size (B)", "Entropy", "Pi", "Pi Deviation (%)", "Chi-Squared"))
+        self.console_out("%30s %15d %15f %15f %15f %15f" % (in_file_name, file_size, entropy, new_pi, pi_deviation, chi_squared))
     
-    def write_entropy_graph(self, in_file_name, out_file_name, step, start=0, end=0):
+    def write_file_graph(self, in_file_name, out_file_name, step, start=0, end=0):
         """
         Calculates entropy oscillations for a file and writes it to CSV spreadsheet.
         """